Cleanup -is flag to -l for SSL.  Another PERL variable name fix.  Clean up debugging options for postmaster and postgres programs.  postmaster -d is no longer optional.  Documentation updates.

+++ src/backend/utils/init/globals.c
@@ -63,7 +63,7 @@ bool		TransactionInitWasProcessed = false;
 
 bool		IsUnderPostmaster = false;
 
-short		DebugLvl = 0;
+int			DebugLvl = 0;
 
 int			DateStyle = USE_POSTGRES_DATES;
 bool		EuroDates = false;
